[traceviewer-libs] Add readme, GitHub workflow and templates,
and expected root folder boilerplate files.

The workflow is a minimal one, that manually builds the two libraries
and runs the unit tests available in traceviewer-react-components.
Since the package.json one level up is not available in the subtree,
we get creative, manually adding the missing dependencies necessary
to sucessfully run the tests.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>