adjust misc configurations
In the repo,so they work with the "subtree" version of traceviwer libraries.
As well in the local version of the subtree, so it integrates correctly in
the repo. e.g. linter configs had relative path that were off one level since
the traceviewer libraries now live one level deeper than when they were local.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>