[traceviewer-base] explicitly request recent prettier dependency
When the traceviewer-libs git subtree was initially added,
running prettier (yarn format:check) would fail locally.
This was since determine to be because this repo here was
already pulling a more recent version of "prettier" than
the theia extension repo: v3.x vs v2.x.

The local version of the subtree already has the libraries'
code adapted to pass the check with prettier@3. That version
was explicitly added as a devDependency to the traceviewer-
react-components library, but for completeness, we now add
it to traceviwewer-base as well, to be sure the format check
will be identical in both repositories, no matter the version
of prettier that may be pulled in the root package.json or
other components, in either / both repositories.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>